This PR fixes a performance bug in DSS: when input arguments are redefined in functions, the compiler can fail to infer them.

Main
```julia
julia> show(stdout, MIME("text/plain"), trial);
BenchmarkTools.Trial: 7203 samples with 1 evaluation.
 Range (min … max):  589.400 μs …   3.344 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     598.700 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   689.941 μs ± 183.943 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █ ▁▄▁▂▂▂▁▂▂▂▂▂▁                                     ▁▁▁▁      ▁
  █████████████████▇▇▇▇▆▅▆▆▅▆▆▅▆▆▆▆▅▆▆▅▆▅▅▅▅▅▅▆▇▅▅▆▆▆▇████▇▇▆▆▆ █
  589 μs        Histogram: log(frequency) by time       1.27 ms <

 Memory estimate: 0 bytes, allocs estimate: 0.
```

This PR:
```julia
BenchmarkTools.Trial: 7986 samples with 1 evaluation.
 Range (min … max):  588.500 μs …  1.658 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     597.800 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   622.034 μs ± 70.136 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▂█  ▁▂▂    ▂          ▁                                      ▁
  ██▆█████▇██████████▇▇▇█▇▆▆▅▅▅▅▅▅▅▅▅▅▄▄▄▄▄▅▃▃▃▃▃▄▃▄▄▃▃▄▂▂▃▃▂▂ █
  588 μs        Histogram: log(frequency) by time       995 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.
```